feat: migrate Announcement component to Typescript
